refactor: change to use loader instead of course fetching
