+++ src/app/hooks/use-course-by-title/utils/is-course.ts
@@ -1,0 +1,7 @@
+import { Course } from '@/app/types';
+
+const isCourse = (obj: object) => {
+  return 'title' in obj && (obj as Course).title != null;
+};
+
+export default isCourse;
