+++ src/features/course-main/courseLoader.ts
@@ -1,0 +1,7 @@
+import { courses } from '@/app/services/data';
+
+const courseLoader = async () => {
+  return courses;
+};
+
+export default courseLoader;
